Add "Select All" command
Rather than adding `DoSelectAll` I've implemented `HandleCommandL` in the views, so that the AppUI can pass off to the view to handle any type of command that it doesn't implement itself. This therfore also does away with `DoCut|Copy|Paste`.

Next up, we should replace `CanCut|Copy|Paste` with `CanHandleCommand`...